% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpseudocolor.R
\name{geom_pseudocolor}
\alias{geom_pseudocolor}
\title{geom_pseudocolor}
\usage{
geom_pseudocolor(
  mapping = NULL,
  data = NULL,
  stat = "pseudocolor",
  position = "identity",
  bins = 5,
  n = 100,
  h = NULL,
  k = 100,
  r = NULL,
  show.legend = T,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{stat}{Four possible options:  \c
pseudocolor: calculates densitites as a continuous spectrum, requires bins, n, hr
pseudocolor_binned: groups densities into bins, requires bins, n, h \cr
pseudocolor_nn: calculates euclidan distance of nearest neighbours, requires k \cr
pseudocolor_nn_count: counts number nearest neighbours, requires k, r}

\item{bins}{Number of bins to split densitities into}

\item{n}{Resolution of the KDE grid}

\item{h}{KDE bandwidth}

\item{k}{Number of nearest neighbours}

\item{r}{Maximum distance to classify as nearest neighbour}
}
\description{
Creates pseudocolor point plot with specified stat calculation
}
