% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpseudocolor.R
\name{scale_color_flowjo_discrete}
\alias{scale_color_flowjo_discrete}
\title{Continuous color palette for flowjo colors}
\usage{
scale_color_flowjo_discrete(...)
}
\description{
Continuous color palette for flowjo colors
}
