% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpseudocolor.R
\docType{data}
\name{StatNN}
\alias{StatNN}
\title{Stat for nearest neighbour distance Pseudocolor Density}
\format{
An object of class \code{StatNN} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatNN
}
\description{
Calculate density based on mean Euclidean distance of nearest neighbours
}
\keyword{internal}
