% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpseudocolor.R
\docType{data}
\name{StatNNCount}
\alias{StatNNCount}
\title{Custom Stat for number of nearest neighbours}
\format{
An object of class \code{StatNNCount} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatNNCount
}
\description{
Calculate number of nearest neighbours within a specified radius
}
\keyword{internal}
