% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpseudocolor.R
\docType{data}
\name{StatPseudcolorBinned}
\alias{StatPseudcolorBinned}
\title{Custom Stat for Binned KDE Pseudocolor Density}
\format{
An object of class \code{StatPseudcolorBinned} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatPseudcolorBinned
}
\description{
Uses 2D kernel density estimation to assign discrete color bins to scatter points.
}
\keyword{internal}
