% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpseudocolor.R
\docType{data}
\name{StatPseudocolor}
\alias{StatPseudocolor}
\title{Custom Stat for Continuous KDE Pseudocolor Density}
\format{
An object of class \code{StatPseudocolor} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatPseudocolor
}
\description{
Uses 2D kernel density estimation to assign continuous color to scatter points.
}
\keyword{internal}
